+++ src/main.tsx
@@ -10,11 +10,13 @@ import { TooltipProvider } from '@/components/ui/tooltip'
 import App from '@/App'
 import '@/index.css'
 
+const basename = import.meta.env.BASE_URL.replace(/\/$/, '')
+
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <ThemeProvider>
       <ApolloProvider client={client}>
-        <BrowserRouter>
+        <BrowserRouter basename={basename || undefined}>
           <TooltipProvider>
             <App />
             <Toaster richColors position="bottom-right" />